Settings updation

+++ app/Http/Controllers/SettingsController.php
@@ -6,9 +6,9 @@ use Illuminate\Http\Request;
 use App\Models\Session;
 use App\Models\Settings;
 use App\Models\SessionTime;
-//use App\Models\IntervalTime;
 use App\Http\Requests\TimeSettingsRegistrationRequest;
 use DateTime;
+use Validator;
 
 class SettingsController extends Controller
 {
@@ -19,18 +19,36 @@ class SettingsController extends Controller
     {
         $noOfDays       = 0;
         $noOfSession    = 0;
-        $noOfInterval   = 0;
+        $sessionTimes   = [];
+        $sessionStatus  = 0;
+
         $settings       = Settings::first();
+        $timeSessions   = SessionTime::where('status', 1)->get();
 
         if(!empty($settings) && !empty($settings->id)) {
-            $noOfDays       = $settings->working_days_in_week;
-            $noOfSession    = $settings->session_per_day;
-            $noOfInterval   = $settings->no_of_intervals_per_day;
+            $noOfDays           = $settings->working_days_in_week;
+            $noOfSession        = $settings->session_per_day;
+            $timetableStatus    = $settings->time_table_status;
         }
+
+        $sessionCount = Session::where('status', 1)->count();
+        if(!empty($sessionCount) && $sessionCount > 0 && $sessionCount == ($noOfSession * $noOfDays)) {
+            $sessionStatus = 1;
+        }
+
+        foreach ($timeSessions as $key => $timeSession) {
+            $sessionTimes [$timeSession->session_index] = [
+                "fromTime"  => DateTime::createFromFormat('H:i:s', $timeSession->from_time)->format('H:i A'),
+                "toTime"    => DateTime::createFromFormat('H:i:s', $timeSession->to_time)->format('H:i A'),
+            ];
+        }
+
         return view('timetable.settings', [
-                'noOfDays'      => $noOfDays,
-                'noOfSession'   => $noOfSession,
-                'noOfInterval'  => $noOfInterval
+                'noOfDays'          => $noOfDays,
+                'noOfSession'       => $noOfSession,
+                'timetableStatus'   => $timetableStatus,
+                'sessionTimes'      => $sessionTimes,
+                'sessionStatus'     => $sessionStatus,
             ]);
     }
 
@@ -42,7 +60,6 @@ class SettingsController extends Controller
         $sessionArray   = [];
         $noOfDays       = $request->get('no_of_days');
         $noOfSession    = $request->get('no_of_session');
-        //$noOfInterval   = $request->get('no_of_intervals_per_day');
 
         $day        = [1 => 'Monday', 2 => 'Tuesday', 3=> 'Wednesday', 4 => 'Thursday', 5 => 'Friday', 6 => 'Saturday'];
         $session    = [1 => 'Mo', 2 => 'Tu', 3 => 'We', 4 => 'Th', 5 => 'Fr', 6 => 'Sa'];
@@ -54,8 +71,6 @@ class SettingsController extends Controller
                     'session_index' => $j,
                     'day_name'      => $day[$i],
                     'session_name'  => ($session[$i]. " - ". $j),
-                    'time_from'     => '09:30:00',
-                    'time_to'       => '16:00:00',
                     'status'        => 1
                 ];
             }
@@ -68,7 +83,7 @@ class SettingsController extends Controller
         $settings = new Settings;
         $settings->working_days_in_week     = $noOfDays;
         $settings->session_per_day          = $noOfSession;
-        //$settings->no_of_intervals_per_day  = $noOfInterval;
+        $settings->time_table_status        = 0;
         $settings->status                   = 1;
         if($settings->save())
         {
@@ -89,47 +104,48 @@ class SettingsController extends Controller
     public function timeSettingsAction(TimeSettingsRegistrationRequest $request)
     {
         $noOfSession        = 0;
-        $noOfInterval       = 0;
         $sessionTimeArr     = [];
         $intervalTimeArr    = [];
+        $validator          = Validator::make([], []);
+        $errorFlag          = 0;
 
+        //session time
         $fromTimeSession    = $request->get('from_time');
         $toTimeSession      = $request->get('to_time');
-        /*$prevSessionIndex   = $request->get('prev_session_index');
-        $intervalFromTime   = $request->get('interval_from_time');
-        $intervalToTime     = $request->get('interval_to_time');*/
         
         $settings       = Settings::where('status', 1)->first();
         if(!empty($settings) && !empty($settings->id)) {
             $noOfSession    = $settings->session_per_day;
-            //$noOfInterval   = $settings->no_of_intervals_per_day;
         }
 
-        for ($i=1; $i <= $noOfSession; $i++) { 
+        for ($i=1; $i <= $noOfSession; $i++) {
+            $fromTime   = DateTime::createFromFormat('H:i A', $fromTimeSession[$i])->format('H:i:s');
+            $toTime     = DateTime::createFromFormat('H:i A', $toTimeSession[$i])->format('H:i:s');
+
+            if($fromTime > $toTime) {
+                //invalid time selection
+                $validator->errors()->add('time_error.'.$i, "Session start time should be less than end time!");
+                $errorFlag = 1;
+            }
+
             $sessionTimeArr[] = [
                 'session_index' => $i,
-                'from_time'     => (DateTime::createFromFormat('H:i A', $fromTimeSession[$i])->format('H:i:s')),
-                'to_time'       => (DateTime::createFromFormat('H:i A', $toTimeSession[$i])->format('H:i:s')),
+                'from_time'     => $fromTime,
+                'to_time'       => $toTime,
                 'status'        => 1
             ];
         }
 
-        /*for($j=1; $j < $noOfInterval; $j++) {
-            $intervalTimeArr[] = [
-                'previous_session_id'   => $prevSessionIndex[$j],
-                'from_time'             => (DateTime::createFromFormat('H:i A', $intervalFromTime[$j])->format('H:i:s')),
-                'to_time'               => (DateTime::createFromFormat('H:i A', $intervalToTime[$j])->format('H:i:s')),
-            ];
-        }*/
+        if($errorFlag != 0) {
+            return redirect()->back()->withInput()->withErrors($validator);
+        }
 
         //delete all existing records
         SessionTime::truncate();
-        //IntervalTime::truncate();
 
         $sessionTimeFlag    = SessionTime::insert($sessionTimeArr);
-        //$intervalTimeFlag   = IntervalTime::insert($intervalTimeArr);
 
-        if($sessionTimeFlag/* && $intervalTimeFlag*/)
+        if($sessionTimeFlag)
         {
             return redirect()->back()->with("message","Time Settings saved successfully")->with("alert-class","alert-success");
         }
